Handle Global Trade Item Number codes
So far there is support for two sub-types of those:
- International Article Numbers (EAN) are opened on Open Food Facts. EANs
exists on all kinds of non-food items too of course, but that's the most
useful Open Data site I am aware of for those.
- International Standard Book Numbers (ISBN) are opened on Wikepedia's
ISBN meta search page.

ISMNs and ISSNs could be handled in the same way in case we find a useful
search site for those.