Set a reasonable default window size
Currently, Qrca defaults to a huge window size.

I feel like this is the sort of small app that would benefit from defaulting to portrait mode, like some simple GNOME apps do.

This specific width shows the Save and Share buttons by default, and this specific height is so the text field is centered.

I'm not sure how Plasma Mobile handles default window sizes. I assume it doesn't and just always maximizes/opens the window in fullscreen? In which case `isMobile` is used here just to get a nice window size when working with `QT_QUICK_CONTROLS_MOBILE=1` on the desktop.

![Screenshot_20241121_114214](/uploads/0424b56e636d662c2ee3cc55e580ea3e/Screenshot_20241121_114214.png)