Change AppData name to match actual name
Right now the name in the AppData file is "Barcode Scanner", and this is
what will be shown in Discover and other app stores.

However once installed, a different name will be shown: Qrca. This is
confusing.

Alleviate this confusion by using the app's actual name in the AppData
file. People should still be able to find it by searching for "barcode"
or "scanner" because those words show up in the .desktop file's
GenericName key, and that gets indexed by AppStream and added to the
list of search keywords.